+++ src/nel/log.cc
@@ -12,9 +12,9 @@ Log &operator<<(Log &outs, char const *v) noexcept
     return outs;
 }
 
-Log &operator<<(Log &outs, size_t const v) noexcept
+Log &operator<<(Log &outs, long unsigned int const v) noexcept
 {
-    fprintf(stderr, "%zu", v);
+    fprintf(stderr, "%lu", v);
     return outs;
 }
